# docs/swagger.yaml
definitions:
  models.Address:
    properties:
      address_id:
        type: integer
      hobli:
        type: string
      lat_coordinate:
        type: string
      long_coordinate:
        type: string
      pincode:
        type: string
      street_address:
        type: string
      taluk:
        $ref: '#/definitions/models.TalukMaster'
      taluk_id:
        type: integer
      village:
        type: string
    type: object
...
  models.CountryMaster:
    properties:
      country_id:
        type: integer
      country_name:
        type: string
    type: object
  models.DistrictMaster:
    properties:
      district_id:
        type: integer
      district_name:
        type: string
      state:
        $ref: '#/definitions/models.StateMaster'
      state_id:
        type: integer
    type: object
  models.Property:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      address_id:
        type: integer
      buildingDetails:
        $ref: '#/definitions/models.PropertyBuildingDetails'
      created_at:
        type: string
      income:
        type: string
      landDetails:
        $ref: '#/definitions/models.PropertyLandDetails'
      media:
        $ref: '#/definitions/models.PropertyMedia'
      original_deed:
        type: string
      ownership:
        $ref: '#/definitions/models.PropertyOwnershipDetails'
      property_id:
        type: integer
      property_name:
        type: string
      property_type_id:
        type: integer
      propertyType:
        $ref: '#/definitions/models.PropertyTypeMaster'
      taxDetails:
        $ref: '#/definitions/models.PropertyTaxDetails'
      user_id:
        type: integer
      value:
        type: string
    type: object
  models.PropertyBuildingDetails:
    properties:
      application_no:
        type: string
      building_details_id:
        type: integer
      built_up_area:
        type: string
      plot_size:
        type: string
      property:
        $ref: '#/definitions/models.Property'
      property_id:
        type: integer
      year_of_construction:
        type: string
    type: object
  models.PropertyLandDetails:
    properties:
      acre:
        type: string
      converted:
        type: string
      ec:
        type: string
      gunte:
        type: string
      karab:
        type: string
      khata_no:
        type: string
      khuskitari:
        type: string
      land_details_id:
        type: integer
      mr_no:
        type: string
      property:
        $ref: '#/definitions/models.Property'
      property_id:
        type: integer
      purpose:
        type: string
      rtc:
        type: string
      sy_no:
        type: string
    type: object
  models.PropertyMedia:
    properties:
      media_id:
        type: integer
      photo_link:
        type: string
      property:
        $ref: '#/definitions/models.Property'
      property_id:
        type: integer
      remarks:
        type: string
      scanned_deed_url:
        type: string
    type: object
  models.PropertyOwnershipDetails:
    properties:
      acquisition_type:
        type: string
      incharge:
        type: string
      ownership_id:
        type: integer
      phone_number:
        type: string
      property:
        $ref: '#/definitions/models.Property'
      property_id:
        type: integer
      received_from:
        type: string
      registration_details:
        type: string
      title:
        type: string
    type: object
  models.PropertyTaxDetails:
    properties:
      property:
        $ref: '#/definitions/models.Property'
      property_id:
        type: integer
      receipt_no:
        type: string
      receipt_url:
        type: string
      tax_amount_paid_current_year:
        type: number
      tax_amount_paid_previous_year:
        type: number
      tax_id:
        type: integer
      tax_paid:
        type: boolean
    type: object
  models.PropertyTypeMaster:
    properties:
      property_type_id:
        type: integer
      property_type_name:
        type: string
    type: object
  models.StateMaster:
    properties:
      country:
        $ref: '#/definitions/models.CountryMaster'
      country_id:
        type: integer
      state_id:
        type: integer
      state_name:
        type: string
    type: object
  models.TalukMaster:
    properties:
      district:
        $ref: '#/definitions/models.DistrictMaster'
      district_id:
        type: integer
      taluk_id:
        type: integer
      taluk_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/properties:
    get:
      description: Retrieve all properties with complete details including address,
        land details, tax details, media, and ownership information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Property'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all properties
      tags:
      - Properties
  /api/v1/properties/{id}/files:
    post:
      consumes:
      - multipart/form-data
      description: Upload ScannedDeed and Receipt files to S3 for an existing property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Scanned Deed PDF/Image
        in: formData
        name: scanned_deed
        type: file
      - description: Tax Receipt PDF/Image
        in: formData
        name: receipt
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Upload binary files for a property (Step 2)
      tags:
      - Properties
...
swagger: "2.0"